Fix critical bug + improvements
- Fix calendar event lookup spinning forever, causing ANR
- Add shared prefs import/export (testing still needed; work in
  progress)
- Polish up dock a little bit more

+++ app/src/main/java/com/inipage/homelylauncher/dock/items/PhoneMappedDockItem.java
@@ -88,7 +88,7 @@ public class PhoneMappedDockItem extends ConfigurableAppBackedDockItem {
     }
 
     @Override
-    public long getBasePriority() {
+    public int getBasePriority() {
         return DockItemPriorities.PRIORITY_CALL.getPriority();
     }
 }
